+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Bob versions changelog
 
 ## [Unreleased]
 #### Added
-
+- SHA512 hash of password with salt 'bob' can be specified instead of password (#304)
 
 #### Changed